refactor: 포트폴리오 페이지 코드 분리
