+++ src/app/portfolio/constants/index.ts
@@ -1,0 +1,4 @@
+import { ChartItemType } from "../types";
+
+export const CHART_ITEMS: ChartItemType[] = ["소비", "금융", "부동산", "투자"];
+export const COLORS = ["#40c057", "#2b7a3e", "#245530", "#868e96"] as const;
