+++ src/app/portfolio/types/index.ts
@@ -1,8 +1,15 @@
+export type ChartItemType = "소비" | "금융" | "부동산" | "투자";
+
 export interface PortfolioData {
-  type: string;
+  type: "AGGRESSIVE" | "PASSIVE" | "NORMAL";
   stocks: number;
   bonds: number;
-  savings: number;
   realEstate: number;
+  savings: number;
   description: string;
+}
+
+export interface ChartDataItem {
+  name: ChartItemType;
+  value: number;
 }
